requirements.txt
librosa
numpy
scikit-learn
matplotlib
soundfile
sounddevice
joblib
spotipy
python-dotenv
scipy
seaborn
pandas
imbalanced-learn